requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
transformers==4.35.2
torch==2.1.1
python-multipart==0.0.6
jinja2==3.1.2
safetensors==0.4.0
tweet-preprocessor==0.6.0
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
optree>=0.13.0
sentencepiece==0.1.99 
hf_xet